email-triage-env/requirements.txt
openenv>=0.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
openai>=1.12.0
pydantic>=2.0.0
pytest>=7.0.0
httpx>=0.27.0
pyyaml>=6.0.0
